Convert timer-cutoff
