--- tests/browser/spa/timer-cutoff.spec.js
@@ -1,2 +1,0 @@
-const matcher = require('../../../tools/jil/util/browser-matcher')
-module.exports = matcher.withFeature('wrappableAddEventListener')
